sqflite implementation in weather app

+++ lib/model/weather_model.dart
@@ -173,11 +173,12 @@ class Main {
 class Wind {
   double speed;
   int deg;
-  double gust;
+  int gust;
 
   Wind({this.speed, this.deg, this.gust});
 
   Wind.fromJson(Map<String, dynamic> json) {
+
     speed = json['speed'];
     deg = json['deg'];
     gust = json['gust'];
